Sliding Inventory panel
-Work In Progress-
Sliding inventory panel.  Slides, but is not set properly.  anchors are
also not set properly.

Inventory appears behind the player.  Because the laver for the canvas
in behind the player layer.  The problem is that the background image is
also in the canvas.  So the layer can not be changed without affecting
where the background is rendered.